requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
Werkzeug==3.0.1
APScheduler==3.10.4
requests==2.31.0
SQLAlchemy==2.0.23
Flask-Migrate
python-dotenv
